potentially fix session getting stuck after ~25 days

+++ AssettoServer/Server/SessionState.cs
@@ -10,7 +10,7 @@ public class SessionState
     public int EndTime { get; set; } // TODO
     public long StartTimeMilliseconds { get; set; }
     public int TimeLeftMilliseconds => Configuration.Infinite ? Configuration.Time * 60_000 : (int)(StartTimeMilliseconds + Configuration.Time * 60_000 - _timeSource.ServerTimeMilliseconds);
-    public int SessionTimeMilliseconds => (int)(_timeSource.ServerTimeMilliseconds - StartTimeMilliseconds);
+    public long SessionTimeMilliseconds => _timeSource.ServerTimeMilliseconds - StartTimeMilliseconds;
     public uint TargetLap { get; set; } = 0;
     public uint LeaderLapCount { get; set; } = 0;
     public bool LeaderHasCompletedLastLap { get; set; } = false;
